moved perform_sync_cql to cass.py from effect_dispatcher
